Bubble Sort method changed from Main.java -> SalesRepresentative.java
