+++ Homeworks/src/main/java/org/cursjava/ObjectContainers_Tema2_Optional/SalesRepresentative.java
@@ -1,6 +1,7 @@
 package org.cursjava.ObjectContainers_Tema2_Optional;
 
-import org.jetbrains.annotations.NotNull;
+
+import java.util.List;
 
 public class SalesRepresentative {
     /**
@@ -40,6 +41,33 @@ public class SalesRepresentative {
     public int getWorthOfSales() {
         return worthOfSales;
     }
+    /**
+     * @BubbleSort - this method is used to sort the element of the above list by @worthOfSales;
+     **/
+    //Creating a method called bubbleSort. As a entry parameter we provide the List of SalesRepresentative, using this method we sort descending the list;
+    public static void bubbleSort(List<SalesRepresentative> list) {
+        int n = list.size();
+        boolean swapped;
+
+        for (int i = 0; i < n - 1; i++) {
+            swapped = false;
+            for (int j = 0; j < n - i - 1; j++) {
+                SalesRepresentative current = list.get(j);
+                SalesRepresentative next = list.get(j + 1);
+
+                // "<" indica daca vrem sa afisam crescator sau descrescator lista;
+                if (current.getWorthOfSales() < next.getWorthOfSales()) {
+                    list.set(j, next);
+                    list.set(j + 1, current);
+                    swapped = true;
+                }
+            }
+            //swapped == false;
+            if (!swapped) {
+                break;
+            }
+        }
 
 
+    }
 }
